Add public chart page with force graph visualization
