+++ package.json
@@ -11,7 +11,8 @@
   "dependencies": {
     "react": "19.1.0",
     "react-dom": "19.1.0",
-    "next": "15.5.4"
+    "next": "15.5.4",
+    "react-force-graph": "file:packages/react-force-graph"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4",
