+++ packages/react-force-graph/index.js
@@ -1,0 +1,1 @@
+export { ForceGraph2D } from "./ForceGraph2D.js";
